requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.50
alembic==1.12.0
pydantic==1.10.12
Jinja2==3.1.2
python-dotenv==1.0.0
APScheduler==3.10.1
email-validator==1.3.1
requests==2.31.0
twilio==8.0.0
pytest==7.4.0
pytest-cov==4.1.0
pymongo==4.5.0
